+++ lib/models/user.js
@@ -18,6 +18,9 @@ module.exports = function (sequelize, DataTypes) {
       type: DataTypes.STRING,
       unique: true
     },
+    name: {
+      type: DataTypes.STRING
+    },
     profile: {
       type: DataTypes.TEXT
     },
@@ -72,7 +75,7 @@ module.exports = function (sequelize, DataTypes) {
         if (!user) {
           return null
         }
-        return user.profile ? User.parseProfile(user.profile) : (user.email ? User.parseProfileByEmail(user.email) : null)
+        return user.profile ? User.parseProfile(user.profile) : (user.email ? User.parseProfileByUser(user) : null)
       },
       parseProfile: function (profile) {
         try {
@@ -143,9 +146,10 @@ module.exports = function (sequelize, DataTypes) {
         }
         return photo
       },
-      parseProfileByEmail: function (email) {
+      parseProfileByUser: function (user) {
+        var email = user.email
         return {
-          name: email.substring(0, email.lastIndexOf('@')),
+          name: user.name || email.substring(0, email.lastIndexOf('@')),
           photo: generateAvatarURL('', email, false),
           biggerphoto: generateAvatarURL('', email, true)
         }
